refactor(nodes): define NodeInfoProvider trait for decoupling
Move NodeInfo, NodeProviderInfo, and NodeInfoProvider trait to
moltis-node-exec-types as the single source of truth. The tools crate
now re-exports these types instead of defining duplicates.

Security note: This PR only adds an interface (Trait). It does not change
any execution logic. This is a prerequisite for safely moving the
remaining provider logic.

Builds on #683 (types & constants) and #685 (helper functions).
