+++ crates/node-exec-types/Cargo.toml
@@ -5,6 +5,8 @@ edition.workspace    = true
 repository.workspace = true
 
 [dependencies]
+anyhow         = { workspace = true }
+async-trait    = { workspace = true }
 moltis-metrics = { optional = true, workspace = true }
 serde          = { workspace = true, features = ["derive"] }
 tracing        = { workspace = true }
